refactor: custom hook
